feat: fractional tokens (0.25) + analytical refinement chat
Backend:
- tokens column migrated from INTEGER to NUMERIC(10,2) via startup migration
- deduct_token() accepts amount param (default 1.0, chat uses 0.25)
- Added logging to deduct_token for debugging
- New POST /analysis/chat endpoint with 0.25 token cost per message
- Chat uses Claude Sonnet 4.5 with full context memory per session

Frontend:
- Token display uses parseFloat (supports 14.75 etc.)
- New s-refine screen: chat UI with IA Analítica by Teo Mariscal
- Button in intervention plan: Afinar plan con IA (−0.25 tok.)
- Accumulated conversation memory within session
- 402 handling → redirects to tokens screen
- Token badge updates after every AI reply

+++ app/models/user.py
@@ -1,6 +1,6 @@
 import uuid
 from datetime import datetime
-from sqlalchemy import Column, String, Integer, DateTime
+from sqlalchemy import Column, String, DateTime, Numeric
 from sqlalchemy.dialects.postgresql import UUID
 from app.database import Base
 
@@ -11,5 +11,5 @@ class User(Base):
     id             = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     email          = Column(String(255), unique=True, nullable=False, index=True)
     password_hash  = Column(String(255), nullable=False)
-    tokens         = Column(Integer, default=5, nullable=False)  # 5 tokens de regalo
+    tokens         = Column(Numeric(10, 2), default=5, nullable=False)  # 5 tokens de regalo
     created_at     = Column(DateTime, default=datetime.utcnow)
